pyproject.toml
[tool.poetry]
name = "crypto-trading-bot"
version = "0.1.0"
description = "Collection of Crypto Trading Bots"
authors = [
    "amazingchow <jianzhou42@163.com>",
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
colorama = "^0.4.6"
httpx = "^0.28.1"
jsonschema = "^4.23.0"
loguru = "^0.7.3"
motor = "^3.6.0"
pandas = "^2.2.3"
python-binance = "^1.0.25"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"
tenacity = "^9.0.0"
ujson = "^5.10.0"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

[tool.sort.config]
# More info: https://pypi.org/project/poetry-sort/
auto = true
case-sensitive = false
sort-python = false
format = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope="function"

[tool.isort]
# More info: https://pycqa.github.io/isort/docs/configuration/action_comments.html
profile = "hug"
src_paths = ["."]
known_local_folder = [
    "internal"
]
sections = ["FUTURE", "STDLIB", "FIRSTPARTY", "THIRDPARTY", "LOCALFOLDER"]
